backend/requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.50
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Werkzeug==2.3.7
email-validator==2.1.0
pytest==7.4.0